updated functions
i separated the loop function to draw the grid from the ask prompt question for grid size.

I will keep the code crossed out for now but I needed to use a simple example to figure out how saving a function as a variable and using that variable as a parameter in the whole function.

I created a whole function to run everything:
1. to ask how big of grid
2. return that value and save it as a parameter
3. use the parameter in grid creation function.